chore: move to table definitions
