+++ packages/shared/src/server/index.ts
@@ -16,7 +16,6 @@ export * from "./auth/invalidateApiKeys";
 export * from "./auth/customSsoProvider";
 export * from "./auth/gitHubEnterpriseProvider";
 export * from "./auth/userProjectRoleAuth";
-export * from "./auth/usersTable";
 export * from "./llm/fetchLLMCompletion";
 export * from "./llm/errors";
 export * from "./llm/utils";
