+++ packages/shared/src/tableDefinitions/index.ts
@@ -6,3 +6,4 @@ export * from "./datasetItemsTable";
 export * from "./mapDatasetRunsTable";
 export * from "./datasetRunsTable";
 export * from "./datasetRunItemsTable";
+export * from "./usersTable";
